update hdb.q
An update to my last pull request. 

Problem: I keep getting "maiNDB: Path of file cannot be found" when hdb.q is called while qExplorer is writing to disk.

Solution: This should prevent the hdb from reloading during the period of time in which qExplorer is writing, without needing to open a connection to qExplorer.